requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.0.0
certifi==2017.1.23
configparser==3.5.0
decorator==4.0.11
entrypoints==0.2.2
enum34==1.1.6
functools32==3.2.3.post2
future==0.16.0
html5lib==0.999999999
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==4.4.1
packaging==16.8
pandocfilters==1.4.1
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-telegram-bot==5.3.0
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.13.0
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.3
traitlets==4.3.2
ujson==1.35
urllib3==1.20
wcwidth==0.1.7
webencodings==0.5
widgetsnbextension==2.0.0
